+++ branches/v0.70/include/tmv/TMV_BandQRD.h
@@ -140,8 +140,8 @@ namespace tmv {
         struct BandQRDiv_Impl;
         std::auto_ptr<BandQRDiv_Impl> pimpl;
 
-        int colsize() const;
-        int rowsize() const;
+        ptrdiff_t colsize() const;
+        ptrdiff_t rowsize() const;
 
     private :
