+++ branches/v0.70/include/tmv/TMV_BandSVD.h
@@ -192,8 +192,8 @@ namespace tmv {
         //
 
         void thresh(TMV_RealType(T) toler, std::ostream* debugout=0) const;
-        void top(int neigen, std::ostream* debugout=0) const;
-        int getKMax() const;
+        void top(ptrdiff_t neigen, std::ostream* debugout=0) const;
+        ptrdiff_t getKMax() const;
 
         //
         // Access Decomposition
@@ -210,8 +210,8 @@ namespace tmv {
         struct BandSVDiv_Impl;
         std::auto_ptr<BandSVDiv_Impl> pimpl;
 
-        int colsize() const;
-        int rowsize() const;
+        ptrdiff_t colsize() const;
+        ptrdiff_t rowsize() const;
 
     private :
